[QSP-1239] Add 0.5 service times per week option

+++ customer_dashboard/views.py
@@ -1192,7 +1192,7 @@ def new_order_4(request):
         context["schedule_window"] = request.GET.get("schedule_window", "")
         context["times_per_week"] = request.GET.get("times_per_week", "")
         if context["times_per_week"]:
-            context["times_per_week"] = int(context["times_per_week"])
+            context["times_per_week"] = Decimal(context["times_per_week"])
         context["shift_count"] = request.GET.get("shift_count", "")
         if context["shift_count"]:
             context["shift_count"] = int(context["shift_count"])
@@ -1617,7 +1617,7 @@ def cart(request):
         waste_type_id = request.POST.get("waste_type")
         schedule_window = request.POST.get("schedule_window", "Morning (7am-11am)")
         times_per_week = (
-            int(request.POST.get("times_per_week"))
+            Decimal(request.POST.get("times_per_week"))
             if request.POST.get("times_per_week")
             else None
         )
